Update sketch_nov29a.ino
Hi Cam,
This script is for running a series of pumps and solenoid valves in the lab. I have the hardware all set up and running fine when just testing isolated components.
The current issue is that the actuators are stuck in the while {} that runs between lines 114-140. Any thoughts on why it is sticking here would be great!
Thanks!